+++ core/src/main/scala/com/jpro/web/sessionmanager/SessionManagerDesktop.scala
@@ -43,7 +43,7 @@ class SessionManagerDesktop(val webApp: WebApp) extends SessionManager { THIS =>
 
         //setView() ???
         webApp.getTransition((THIS.view,view,!pushState)).doTransition(webApp,THIS.view,view)
-        if(THIS.view != null) {
+        if(THIS.view != null && THIS.view != view) {
           THIS.view.onClose()
           THIS.view.sessionManager = null
           markViewCollectable(THIS.view)